+++ src/components/atoms/file-upload-field/index.tsx
@@ -5,18 +5,25 @@ type FileUploadFieldProps = {
   onFileChosen: (files: any[]) => void
   filetypes: string[]
   errorMessage?: string
-  placeholder?: string
+  placeholder?: React.ReactElement | string
   className?: string
-  text?: string
+  text?: React.ReactElement | string
 }
+
+const defaultText = (
+  <span>
+    Drop your images here, or{" "}
+    <span className="text-violet-60">click to browse</span>
+  </span>
+)
 
 const FileUploadField: React.FC<FileUploadFieldProps> = ({
   onFileChosen,
   filetypes,
-  placeholder,
   errorMessage,
   className,
-  text = "Drop your images here, or",
+  text = defaultText,
+  placeholder = "",
 }) => {
   const inputRef = useRef(null)
   const [fileUploadError, setFileUploadError] = useState(false)
@@ -69,9 +76,7 @@ const FileUploadField: React.FC<FileUploadFieldProps> = ({
       )}
     >
       <div className="flex flex-col items-center">
-        <p>
-          {text} <span className="text-violet-60">click to browse</span>
-        </p>
+        <p>{text}</p>
         {placeholder}
       </div>
       {fileUploadError && (
