+++ src/components/templates/add-products-modal/index.tsx
@@ -1,9 +1,9 @@
 import { Product } from "@medusajs/medusa"
 import { useAdminProducts } from "medusa-react"
 import * as React from "react"
-import Button from "../../../components/fundamentals/button"
-import Modal from "../../../components/molecules/modal"
-import { SelectableTable } from "../../../components/templates/selectable-table"
+import Button from "../../fundamentals/button"
+import Modal from "../../molecules/modal"
+import { SelectableTable } from "../selectable-table"
 import useQueryFilters from "../../../hooks/use-query-filters"
 import { columns, ProductHeader, ProductRow } from "./product-table-config"
 import { refreshItems } from "./utils"
